+++ src/main/java/mods/eln/gridnode/transformer/GridTransformerRender.kt
@@ -10,20 +10,20 @@ import java.io.DataInputStream
 
 class GridTransformerRender(entity: TransparentNodeEntity, descriptor: TransparentNodeDescriptor) : GridRender(entity, descriptor) {
     val desc = descriptor as GridTransformerDescriptor
-    private var load = SlewLimiter(0.5f)
+    private var load = SlewLimiter(0.5)
 
     init {
         addLoopedSound(object : LoopedSound("eln:Transformer", coordonate(), ISound.AttenuationType.LINEAR) {
-            override fun getVolume() = Math.max(0f, (load.position - desc.minimalLoadToHum) / (1 - desc.minimalLoadToHum))
+            override fun getVolume() = Math.max(0.0, (load.position - desc.minimalLoadToHum) / (1 - desc.minimalLoadToHum)).toFloat()
         })
     }
 
     override fun networkUnserialize(stream: DataInputStream) {
         super.networkUnserialize(stream)
-        load.target = stream.readFloat()
+        load.target = stream.readDouble()
     }
 
-    override fun refresh(deltaT: Float) {
+    override fun refresh(deltaT: Double) {
         super.refresh(deltaT)
         load.step(deltaT)
     }
