+++ src/main/java/mods/eln/sixnode/electricalgatesource/ElectricalGateSourceRender.java
@@ -44,7 +44,7 @@ public class ElectricalGateSourceRender extends SixNodeElementRender {
         } else {
             LRDU.Down.glRotateOnX();
         }
-        descriptor.draw(interpolator.get(), UtilsClient.distanceFromClientPlayer(this.tileEntity), tileEntity);
+        descriptor.draw((float)interpolator.get(), UtilsClient.distanceFromClientPlayer(this.tileEntity), tileEntity);
     }
 
     @Override
