+++ src/main/java/mods/eln/sound/LoopedSoundManager.kt
@@ -3,8 +3,8 @@ package mods.eln.sound
 import mods.eln.Eln
 import net.minecraft.client.Minecraft
 
-class LoopedSoundManager(val updateInterval: Float = 0.5f) {
-    private var remaining = 0f
+class LoopedSoundManager(val updateInterval: Double = 0.5) {
+    private var remaining = 0.0
     private val loops = mutableSetOf<LoopedSound>()
 
     fun add(loop: LoopedSound?) {
@@ -18,7 +18,7 @@ class LoopedSoundManager(val updateInterval: Float = 0.5f) {
     // takes in two points and gets the squared distance delta between them
     inline fun sqDistDelta(cx: Double, cy: Double, cz: Double, px: Double, py: Double, pz: Double) = (cx - px) * (cx - px) + (cy - py) * (cy - py) + (cz - pz) * (cz - pz)
 
-    fun process(deltaT: Float) {
+    fun process(deltaT: Double) {
         remaining -= deltaT
         if (remaining <= 0) {
             val soundHandler = Minecraft.getMinecraft().soundHandler
