+++ src/main/java/mods/eln/transparentnode/powercapacitor/PowerCapacitorRender.java
@@ -33,7 +33,7 @@ public class PowerCapacitorRender extends TransparentNodeElementRender {
     }
 
     @Override
-    public void refresh(float deltaT) {
+    public void refresh(double deltaT) {
 
     }
